requirements.txt
flask
azure-storage-blob
